KAFKA-1866 workaround: remove metrics when a NoSuchElementException is discovered when reading a Gauge
With Kafka  <= 0.8.2.2 there is an issue if topics get deleted or partions are moved between brokers.
The metrics are not get deleted in this case and because they are implemented as a Gauge, a NoSuchElementException
is thrown when the metrics are reported.
There is already a fix for this, see https://issues.apache.org/jira/browse/KAFKA-1866 but it did not make
it into an 0.8.x release. Because of this we implemented a workaround for this within the `FilterMetricsPredicate`.